Front end logic + backend integration [FIT-71]
Automatically populate empty exercise log with exercises from the workout plan.
Delete plans from database once they are completed.

+++ frontend/view/ExerciseLog/exerciseLogView.js
@@ -2,9 +2,11 @@ import * as React from 'react';
 import { ActivityIndicator, FlatList, Text, View, TouchableOpacity } from 'react-native';
 import { styles } from '../../style/styles';
 import { cleanString, numberToTime } from '../../utility/format.js';
-import { getExerciseSets, deleteExerciseSet } from '../../controller/Exercise/exerciseLogController' 
+import { getExerciseSets, deleteExerciseSet, getExerciseById } from '../../controller/Exercise/exerciseLogController' 
 import { postUserActivity, getUserActivity, patchUserActivity } from '../../controller/UserActivity/userActivityController'
-import { retrieveUserId } from '../../utility/dataHandler.js'
+import { retrievePlanId, retrieveUserId } from '../../utility/dataHandler.js'
+import { getWorkoutPlan, patchWorkoutPlan, patchUser, deleteWorkoutPlan } from '../../controller/Exercise/workoutPlanController';
+import { postSet } from '../../controller/Exercise/exerciseRecorderController';
 
 export function ExerciseLog({ navigation, route }) {
   /* Create hooks */
@@ -12,14 +14,66 @@ export function ExerciseLog({ navigation, route }) {
   const [data, setData] = React.useState(null);
   const [activityData, setActivityData] = React.useState(null);
   const [logMode, setLogMode] = React.useState('select');
+  const [hasCompleted, setHasCompleted] = React.useState(false);
 
   const { date } = route.params;
+
+  // Take an exercise from the plan and turn it into a set
+  const getSetFromExercise = async (id, index) => {
+    const json = await getExerciseById(id);
+    const userId = await retrieveUserId();
+    const exerciseName = cleanString(json.data.ExerciseName);
+    const isCardio = cleanString(json.data.MuscleGroup) === 'Cardio' ? true : false;
+
+    const body = {
+      _id: index, 
+      userId: userId, 
+      exercise_name: exerciseName,
+      is_cardio: isCardio,
+      first_value: 0,
+      second_value: 0,
+      date: date,
+      completed: false
+    };
+
+    return body;
+  }
+
+  // Get workout data from exercise plan
+  const getWorkoutData = async (userPlanId) => {
+    try {
+      const json = await getWorkoutPlan(cleanString(userPlanId));
+      const workoutCounter = json.data.workout_counter;
+      const workoutFrequency = json.data.workouts.length;
+      const workout = json.data.workouts[workoutCounter % workoutFrequency];
+      const workoutData = [];
+      
+      for (let i = 0; i < workout.length; i++){
+        workoutData.push(await getSetFromExercise(workout[i], i));
+      }
+      return workoutData;
+    } catch (error) {
+      console.error(error);
+    }
+  }
 
   /* Pull user's sets */
   const getSets = async () => {
     try {
       const json = await getExerciseSets(cleanString(date));
-      setData(json.data);
+      if (json.data.length == 0){
+        // Need to check if user has a plan in place
+        const userPlanId = await retrievePlanId();
+        if (userPlanId == 'null'){
+          setData(json.data);
+        }
+        else {
+          setData(await getWorkoutData(userPlanId));
+        }
+      }
+      else {
+        setData(json.data);
+      }
     } catch (error) {
       console.error(error);
     } finally {
@@ -41,7 +95,6 @@ export function ExerciseLog({ navigation, route }) {
   const deleteSets = async (id) => {
     try {
       const json = await deleteExerciseSet(id);
-      getSets();
     } catch (error) {
       console.error(error);
     }
@@ -73,15 +126,61 @@ export function ExerciseLog({ navigation, route }) {
     const json = await patchUserActivity(id, body);
   }
 
+  const savePlannedExercises = async () => {
+    // get user Id
+    const userId = await retrieveUserId();
+
+    // Loop through all exercises and save them in database
+    for (let i = 0; i < data.length; i++){
+      const body = JSON.stringify({
+        userId: userId, 
+        exercise_name: data[i].exercise_name,
+        is_cardio: data[i].is_cardio,
+        first_value: data[i].first_value,
+        second_value: data[i].second_value,
+        date: date
+      });
+      const json = await postSet(body);
+    }
+  }
+
+  const incrementWorkoutCounter = async () => {
+    // get necessary workoutplan data
+    const planId = await retrievePlanId();
+    const json = await getWorkoutPlan(cleanString(planId));
+    let workout_counter = json.data.workout_counter + 1;
+    const workout_frequency = json.data.workouts.length;
+    const workout_weeks = json.data.workout_length;
+
+    // True if user has completed all perscribed workouts
+    if (workout_counter >= workout_frequency * workout_weeks){
+      // remove planId from user object
+      const body = JSON.stringify({
+        workout_plan : null,
+      });
+      patchUser(body);
+
+      // Delete workout plan from database
+      deleteWorkoutPlan(planId)
+    }
+    else {
+      // Update workout counter
+      const body = JSON.stringify({
+        workout_counter : cleanString(workout_counter),
+      });
+      patchWorkoutPlan(body);
+    }
+  }
+
   /* Perform update/delete depending on queued action */
   const doEvent = (item) => {
-    if (logMode === 'delete'){
+    if (logMode === 'delete' && item.completed == true){
       setActivityData(null);
       Promise.resolve(deleteSets(cleanString(item._id)))
       .then(() => getSets())
       .then(() => getActivity());
     }
-    else if (logMode === 'update'){
+    else if (logMode === 'update' && item.completed == true){
       navigation.navigate('Record Exercise', {
         exercise_name: item.exercise_name,
         exercise_group: item.is_cardio ? 'Cardio' : 'Not Cardio',
@@ -91,9 +190,17 @@ export function ExerciseLog({ navigation, route }) {
         date: item.date,
       })
     }
-    else if (logMode === 'complete')
+    else if (logMode === 'complete' && item.completed == false)
     {
       item.completed = true;
+      // Need to check if this is the first planned exercise is completed. If so,
+      // we need to save this exercises along with all the others and increment
+      // the workout counter.
+      if (!hasCompleted){
+        savePlannedExercises();
+        incrementWorkoutCounter();
+        setHasCompleted(true);
+      }
     }
     setLogMode('select')
   }
